pyproject.toml
[project]
name = "hex_dora_node_mujoco"
version = "0.0.1-alpha.2"
authors = [{ name = "Dong Zhaorui", email = "847235539@qq.com" }]
description = "Dora Node for using Hex Mujoco"
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "dora-rs>=0.3.9",
    "hex-driver-mujoco>=0.0.0,<0.1.0",
    "hex-util-robot>=0.0.0,<0.1.0",
]

[dependency-groups]
dev = ["pytest>=8.1.1", "ruff>=0.9.1"]

[project.scripts]
hex-dora-mujoco-archer-y6 = "hex_dora_node_mujoco.main_archer_y6:main"
hex-dora-mujoco-e3-desktop = "hex_dora_node_mujoco.main_e3_desktop:main"
hex-dora-mujoco-test-archer-y6 = "hex_dora_node_mujoco.main_archer_y6_test:main"
hex-dora-mujoco-test-e3-desktop = "hex_dora_node_mujoco.main_e3_desktop_test:main"

[tool.ruff.lint]
extend-select = [
    "D",    # pydocstyle
    "UP",   # Ruff's UP rule
    "PERF", # Ruff's PERF rule
    "RET",  # Ruff's RET rule
    "RSE",  # Ruff's RSE rule
    "NPY",  # Ruff's NPY rule
    "N",    # Ruff's N rule
    "I",    # Ruff's I rule
]
